First refactoring for issue #54
