+++ de.variantsync.core/src/main/java/de/variantsync/core/marker/AMarkerInformation.java
@@ -1,4 +1,4 @@
-package de.tubs.variantsync.core.utilities;
+package de.variantsync.core.marker;
 
 import java.util.UUID;
